Moved messageList component and functionality from dashboard page to separate message board page
