+++ kyue-studio-backend/data/contact_form_messages.json
@@ -24,16 +24,6 @@
     {
       "firstName": "1234",
       "lastName": "1234",
-      "email": "as@23424.com",
-      "subject": "35",
-      "tags": [
-        "Job Opportunity"
-      ],
-      "message": "sdlfjs"
-    },
-    {
-      "firstName": "1234",
-      "lastName": "1234",
       "email": "3242436fs344@23424.com",
       "subject": "35",
       "tags": [
